requirements.txt
# Telegram Bot 框架
python-telegram-bot==20.7

# 数据库相关
# aiosqlite==0.19.0  # 注释掉旧的SQLite依赖
supabase==2.3.4
asyncpg==0.29.0

# Web框架
flask==3.0.0
aiohttp==3.9.1
aiofiles==23.2.0

# HTTP客户端
httpx==0.25.2
requests==2.31.0

# 数据处理
pydantic==2.5.0

# 日志
structlog==23.2.0

# 环境变量
python-dotenv==1.0.0

# 图像处理
Pillow==10.1.0

# 加密
cryptography==41.0.8

# 时间处理
python-dateutil==2.8.2

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0 